Get rid of CocoaInput dependency.
