+++ build.gradle
@@ -2,18 +2,13 @@ buildscript {
     repositories {
         jcenter()
         maven { url = "https://files.minecraftforge.net/maven" }
-        maven {
-            url "https://plugins.gradle.org/m2/"
-        }
     }
     dependencies {
         classpath 'net.minecraftforge.gradle:ForgeGradle:2.3-SNAPSHOT'
-        classpath "com.wynprice.cursemaven:CurseMaven:2.1.1"
     }
 }
 
 apply plugin: 'net.minecraftforge.gradle.forge'
-apply plugin: "com.wynprice.cursemaven"
 
 version = "1.1.1"
 group = "cn.yesterday17"
@@ -32,14 +27,10 @@ minecraft {
     replace '@VERSION@', project.version
 }
 
-dependencies {
-    compile "curse.maven:cocoainput:2540534"
-}
-
 jar {
     manifest {
         attributes([
-                "TweakClass": "cn.yesterday17.kokoalinux.lwjgl.Tweaker",
+                "TweakClass": "cn.yesterday17.kokoalinux.tweaker.KokoaTweaker",
         ])
     }
