+++ libkokoa/library.c
@@ -1,108 +1,50 @@
 #include "library.h"
 
-void (*callbackDone)();
+int *(*getDisplayPosition)();
 
-int *(*callbackDraw)(int, int, int, short, int, char *, wchar_t *, int, int, int);
-
-void setCallback(int *(*c_draw)(int, int, int, short, int, char *, wchar_t *, int, int, int), void(*c_done)()) {
-    callbackDraw = c_draw;
-    callbackDone = c_done;
+void emptyCallback(XIC xic, XPointer clientData, XPointer data) {
 }
 
-void preeditCalet(XIC xic, XPointer clientData, XPointer data) {
-
-}
-
-void preeditStart(XIC xic, XPointer clientData, XPointer data) {
-
-}
-
-void preeditDone(XIC xic, XPointer clientData, XPointer data) {
-    callbackDone();
-}
-
-void preeditDraw(XIC xic, XPointer clientData, XIMPreeditDrawCallbackStruct *structure) {
-    int *array;
-    int secondary = 0;
-    int length = 0;
-    if (structure->text) {
-        int i = 0;
-        int secondary_determined = 0;
-        for (i = 0; i != structure->text->length; i++) {
-            if (!secondary_determined && (structure->text->feedback[i] & XIMSecondary) != 0) {
-                secondary = i;
-                secondary_determined = 1;
-            }
-            if (secondary_determined &&
-                (structure->text->feedback[i] == 0 || (structure->text->feedback[i] & XIMSecondary) != 0)) {
-                length++;
-            } else if (secondary_determined)break;
-        }
-    }
-
-    if (structure->text != NULL) {
-        array = callbackDraw(
-                structure->caret,
-                structure->chg_first,
-                structure->chg_length,
-                structure->text->length,
-                structure->text->encoding_is_wchar,
-                structure->text->encoding_is_wchar ? "" : structure->text->string.multi_byte,
-                structure->text->encoding_is_wchar ? structure->text->string.wide_char : L"",
-                0,
-                secondary,
-                secondary + length
-        );
-    } else {
-        array = callbackDraw(
-                structure->caret,
-                structure->chg_first,
-                structure->chg_length,
-                0,
-                0,
-                "",
-                L"",
-                0,
-                0,
-                0
-        );
-    }
+void preeditDraw(XIC xic, XPointer clientData, XIMPreeditDrawCallbackStruct *s) {
+    int *array = getDisplayPosition();
 
     fflush(stdout);
-    XVaNestedList attr;
     XPoint place;
     place.x = array[0];
     place.y = array[1];
-    attr = XVaCreateNestedList(0, XNSpotLocation, &place, NULL);
+
+    XVaNestedList attr = XVaCreateNestedList(0, XNSpotLocation, &place, NULL);;
     XSetICValues(xic, XNPreeditAttributes, attr, NULL);
     XFree(attr);
 }
 
+void setEmpty(XICCallback *cb) {
+    cb->client_data = NULL;
+    cb->callback = emptyCallback;
+}
 
-XICCallback calet, start, done, draw;
+XICCallback empty, draw;
 
 XVaNestedList preeditCallbacksList() {
-    start.client_data = NULL;
-    start.callback = preeditStart;
+    empty.client_data = NULL;
+    empty.callback = emptyCallback;
 
-    done.client_data = NULL;
-    done.callback = preeditDone;
-
-    draw.client_data = NULL;
+    empty.client_data = NULL;
     draw.callback = preeditDraw;
-
-    calet.client_data = NULL;
-    calet.callback = preeditCalet;
     return XVaCreateNestedList(0, // DUMMY
                                XNPreeditStartCallback,
-                               &start,
+                               &empty,
                                XNPreeditDoneCallback,
-                               &done,
+                               &empty,
                                XNPreeditDrawCallback,
                                &draw,
                                XNPreeditCaretCallback,
-                               &calet,
+                               &empty,
                                NULL); // FINAL
+}
+
+void setDisplayPositionCallback(int *(*c_draw)()) {
+    getDisplayPosition = c_draw;
 }
 
 long createInactiveIC(long xim, long window) {
