+++ src/main/java/cn/yesterday17/kokoalinux/KokoaLinux.java
@@ -1,21 +1,22 @@
 package cn.yesterday17.kokoalinux;
 
-import com.Axeryok.CocoaInput.CocoaInput;
+import cn.yesterday17.kokoalinux.display.DisplayHelper;
+import cn.yesterday17.kokoalinux.gui.GuiChange;
+import cn.yesterday17.kokoalinux.input.InputHelper;
+import cn.yesterday17.kokoalinux.input.InputNative;
 import com.google.common.eventbus.EventBus;
 import com.google.common.eventbus.Subscribe;
+import com.sun.jna.Memory;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.DummyModContainer;
 import net.minecraftforge.fml.common.LoadController;
 import net.minecraftforge.fml.common.ModMetadata;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import net.minecraftforge.fml.common.versioning.ArtifactVersion;
-import net.minecraftforge.fml.common.versioning.VersionParser;
 import org.apache.logging.log4j.Logger;
+import org.lwjgl.opengl.Display;
 
-import java.io.IOException;
 import java.lang.reflect.Field;
-import java.util.Collections;
-import java.util.List;
 
 public class KokoaLinux extends DummyModContainer {
     private static final String MOD_ID = "kokoalinux";
@@ -37,11 +38,6 @@ public class KokoaLinux extends DummyModContainer {
     }
 
     @Override
-    public List<ArtifactVersion> getDependencies() {
-        return Collections.singletonList(VersionParser.parseVersionReference("CocoaInput@[3.1.0,)"));
-    }
-
-    @Override
     public boolean registerBus(EventBus bus, LoadController controller) {
         bus.register(this);
         return true;
@@ -55,13 +51,35 @@ public class KokoaLinux extends DummyModContainer {
 
     @Subscribe
     public void init(FMLInitializationEvent event) {
+        // This prepares for the position to display IME
+        InputNative.instance.setDisplayPositionCallback(
+                () -> {
+                    int[] point = {Display.getX(), Display.getY()};
+                    Memory memory = new Memory(8L);
+                    memory.write(0L, point, 0, 2);
+                    return memory;
+                }
+        );
+
+        // In fact it's public after tweaked
         try {
-            Field f = CocoaInput.class.getDeclaredField("instance");
+            Field f = Class.forName("org.lwjgl.opengl.LinuxEvent").getDeclaredField("enableIME");
             f.setAccessible(true);
-            CocoaInput instance = (CocoaInput) f.get(null);
-            instance.applyController(new LinuxController());
-        } catch (NoSuchFieldException | IOException | IllegalAccessException e) {
+            f.set(null, true);
+        } catch (NoSuchFieldException | IllegalAccessException | ClassNotFoundException e) {
             e.printStackTrace();
         }
+
+        // Destroy original IC & IMC
+        DisplayHelper.destroyLWJGLIC();
+        DisplayHelper.closeLWJGLIM();
+
+        // Construct IM-friendly environment
+        InputHelper.prepareLocale();
+        InputHelper.openIM();
+        InputHelper.createIC(false);
+
+        // For subscribe events
+        MinecraftForge.EVENT_BUS.register(GuiChange.class);
     }
 }
